updated all api routes to version 2
