+++ components/PlotIdPage/PaymentPlanTable.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { Table, Text, Title, Card, Grid, Divider } from "@mantine/core";
-import { PlotDetail } from "@/pages/api/plot/[id]";
+import { PlotDetail } from "@/pages/api/housingScheme/[housingSchemeId]/plot/[id]";
 
 export type PaymentPlanTable = {
   tableRows?: PlotDetail["payment_plan"];
