+++ pages/api/housingScheme/[housingSchemeId]/index.ts
@@ -1,27 +1,29 @@
-// Next.js API route support: https://nextjs.org/docs/api-routes/introduction
 import type { NextApiRequest, NextApiResponse } from "next";
-
-import { prisma } from "../../../db/prisma";
-import { ReturnError } from "../customer/all";
-
-export interface PlotsSelectFields {
-  id: number;
-  dimension: string | null;
-  square_feet: number | null;
-  plot_status: string;
-}
+import { prisma } from "@/db/prisma";
+import { ReturnError } from "@/pages/api/customers";
+import { PlotsSelectFields } from "..";
 
 export default async function allPlots(
   req: NextApiRequest,
   res: NextApiResponse<PlotsSelectFields[] | ReturnError>
 ) {
   try {
+    const { housingSchemeId } = req.query;
+    if (typeof housingSchemeId !== "string") {
+      return res.status(404).json({
+        error: "incorrect query parameter, expect query parameter as string",
+      });
+    }
+    const id = housingSchemeId as string;
     const data = await prisma.plot.findMany({
       select: {
         id: true,
         dimension: true,
         square_feet: true,
         plot_status: true,
+      },
+      where: {
+        housing_scheme: parseInt(id),
       },
     });
     res.status(200).json(data);
