+++ pages/api/payments/delete.ts
@@ -1,8 +1,7 @@
 // Next.js API route support: https://nextjs.org/docs/api-routes/introduction
 import type { NextApiRequest, NextApiResponse } from "next";
 
-import { prisma } from "../../../db/prisma";
-//import { TableRowItem } from "@/components/PlotIdPage/AddPaymentForm/PaymentInputTable";
+import { prisma } from "@/db/prisma";
 
 export interface PostReturnType {
   deleted?: true;
