+++ pages/api/user/addUser.ts
@@ -1,7 +1,5 @@
 import type { NextApiRequest, NextApiResponse } from "next";
-
-import { prisma } from "../../../db/prisma";
-import { TableRowItem } from "@/components/PlotIdPage/AddPaymentForm/PaymentInputTable";
+import { prisma } from "@/db/prisma";
 
 export interface PostReturnType {
   created?: true;
